+++ qtNanomite/qtDLGWindowView.cpp
@@ -42,7 +42,7 @@ qtDLGWindowView::qtDLGWindowView(QWidget *parent, Qt::WFlags flags,qint32 iPID)
 	_iForEntry = 0;
 	_iForEnd = myMainWindow->coreDebugger->PIDs.size();
 
-	for(int i = 0; i < myMainWindow->coreDebugger->PIDs.size(); i++)
+	for(size_t i = 0; i < myMainWindow->coreDebugger->PIDs.size(); i++)
 	{
 		if(myMainWindow->coreDebugger->PIDs[i].dwPID == _iPID)
 			_iForEntry = i; _iForEnd = i +1;
@@ -62,7 +62,7 @@ void qtDLGWindowView::EnumWindow()
 {
 	tblWindowView->setRowCount(0);
 
-	for(int i = _iForEntry; i < _iForEnd;i++)
+	for(size_t i = _iForEntry; i < _iForEnd;i++)
 	{
 		EnumWindows((WNDENUMPROC)clsCallbacks::EnumWindowCallBack,(LPARAM)myMainWindow->coreDebugger->PIDs[i].dwPID);
 	}
